chat with multiple PDFs/requirements.txt
streamlit
# google-genrativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community